requirements.txt
Flask==2.3.2
Flask-WTF-1.1.1 WTForms==3.0.1